app: fix home link highlight (hnt-1980) (#3017)
from the current URL pattern this wouldn't work without faking it
(`/t/:town/channel` is not underneath `/t/:town/home` because of the
`/home` segment). Let me know if I'm missing something but I feel the
more consistency we can have between paths desktop and mobile the
better. In this PR we remove the `home` and switch dynamically between
two exclusive routes:
- desktop `index` (placeholder for redirects and possible a landing page
in the future)
-  mobile `index` (home screen).