app: fix autofocus method (hnt-2023) (#3035)
The autofocus occurring when writing a new message got broken in the
last refactor.

This PR fixes it and simplifies the API for the `focusItem` prop
- `align`: `start |end` 
- `force`: if this flag isn't set, the message will only get focused if
you haven't started scrolling yet (e.g. focus on last message or
highlighted thread when entering a channel)

### misc:
moved some types to a shared file to avoid dependency cycle 

### todo:
I've added a TODO. Currently to detect if the message has been sent by
the same user we're only looking at the timestamp and the userId. This
is a bit naive and I want to give this a though before treating the
issue in an appropriate way.