fix bug in megolm (#3082)
There was a bug in megolm.ts found from running
`sendThreadedMessageHooks.test.tsx` off
[this](jt/hnt-1975-timeline-decrypted-events) timeline branch, whereby 2
consecutive sendMessages we're dispatched to different channels the
cached session would be used to encrypt, which would be the wrong
session for the second encryption event. Fixing this bug by keeping
track of channelId along with the session.