app: fix invites in TouchHome (#3080)
TouchHome didn't have any invite logic, so I made a component that wraps
the content of TouchHome and SpaceHome to account for invite links.
Since we're splitting routes between desktop and mobile, and these two
route components are the entry point for the town, let's be extra
conscious of how changes in one carry over to the other. Added some
tests for TouchHome, so feel free to add anything there.

We're all touching these files, I'll hold off on merging until Theo's
bigger refactor is in.