Update the zionEventHandlers test with permissions (#3140)
Not sure if this is what’s causing the tests to fail or not, but this is
the only difference I can see…

We keep getting
```

use-zion-client:test:dendrite:e2e:   ● Zion event handlers test › onCreateSpace
use-zion-client:test:dendrite:e2e:
use-zion-client:test:dendrite:e2e:     M_UNKNOWN_TOKEN: MatrixError: [401] Unknown token
```

on just this test. I tried retrying the space creation, but the test ran
for 71 seconds on the last run, so i don’t think that’s it.

Should we allow creating a space with tokens but an empty permissions
array?