Delete some unnecessary bits in streamview (#3060)
I’m trying to refactor this to add snapshots and blocks, noticed that
this was pretty weird. Why re-create the type if it’s already the type

As for getToDeviceWirePayloadContent… I’m hoping this was just something
copied from matrix? Not sure why we can’t just use the data the same way
we stuck it into the payload, as it’s typed in protobuf