Handle Miniblock events in the client (#3133)
I needed to add an empty data to the miniblock header content to get
appease the client code. The way the data is structured, I think this is
the right thing to do, even though the client should be able to figure
it out. If it becomes a problem we can fix the client and delete the
empty payload.