Use logNever instead of checkNever when parsing protobufs (#3205)
Use this function in the default case of a exhaustive switch statement
to ensure that all cases are handled, but you don't want to throw an
error.
Typical place you wouldn't want to throw an error - when parsing a
protobuf message on the client. The protocol may have been updated on
the server, but the client hasn't been updated yet. In this case, the
client will receive a case that they can't handle, but it shouldn't
break other messages in the stream. If you throw in the middle of a loop
processing events,then lots of messages will appear lost, when you could
have just gracefully handled a new case.