Initialize streams from snapshots and blocks (#3208)
instead of from list of events starting at inception

This isn’t completely done, because the backend is still sending all
events instead of just the minipool events in the StreamAndCookie, but
all tests are passing and it’s a pretty big chunk to review.

Streams now have three mutators

initialize
prependEvent
appendEvent

initialize sets up state from a snapshot
prepend ignores anything that was already in the snapshot, processes
ephemeral things like channel messages
append does what it always did

fun tidbits
- prepend will notice if it’s the first ever event and add that event’s
hash to the leaf hashes, so that you can post new events to the stream
- initializations are very noisy, for example: they dispatch events for
every membership. We can probably optimize this later to have a smarter
initialized event? the tests were counting on the membership events and
what not so i left it