continuous deploys for river nodes (#3241)
This PR introduces a new github workflow even listener: `on
repository_dispatch`. This is a special trigger that can run any
workflow programmatically. With this addition, the river-deploy script
can run programmatically, on top of the manual trigger. The second half
of the PR makes us of this programmatic trigger, so that whenever we
build the docker image and publish it, we also deploy it to test-beta.