Expose Wallet Link contract through River RPC (#3349)
WalletLink contract is a top-chain contract for storing wallet <-> RK
link information
The following new RPC method are added:
* LinkWallet(...) - linking wallet to RK
* GetWalletLink(...) - getting list of wallets by RK id

At the moment the contract assumes the caller's wallet is getting linked
to the root key. This will be addressed in
https://linear.app/hnt-labs/issue/HNT-2344/extend-walletlink-contract-to-allow-linking-using-node-caller.
Once it is done, the disabled check in the test can be enabled.