app: in CreateSpaceFormStore store tokens as TokenDataStruct (#3370)
Changes the tokens stored in the space creation form store from
`string[]` to `TokenDataStruct[]`, an object with a `contractAddress`
property (to include tokenId, etc in future).

The TokenList component is shared across create space form and the
settings page, hence the file changes outside of create space feature.

UI and functionality wise, nothing changes. This PR simply prepares the
app to be able to submit the gated tokens as objects with various
properties, so that different token types can be used for gating. Lib
work to submit these values to contract as `ExternalTokenStruct` will be
in separate PR.

Tested:
- creating a token gated space
- updating roles/settings