Add streamId as param to CreateStreamRequest (#3380)
Problem: we’d like to be able to forward events to the right node at the
request level AND we don’t want to unpack event envalopes twice.

Solution: all of the rpc request objects contain a stream id at the top
level except for CreateStream. In this PR I’m adding that parameter so
that we can properly route the request without inspecting the inception
event.