Prevent cleaning DB on node restart (#3435)
In HNT-2486 a bug is described that could only happen if the db schema
is wiped on node restart. This seems to be the line that is cleaning the
DB schema with a CASCADE clause to remove all tables therein. In local
development and CI, setting `SKIP_GENKEY=false` configuration parameter
has the effect of generating a new db on node start and should be the
preferred way to point to a new backend store for the node.