app: remove uses of useMatrixCredentials (#3423)
In favor of useAuth (which uses useConnectivity, which is dependent on
correct primaryProtocol), and useMyProfile.

Left this and useCasablancaCredentials exported in lib b/c sample-app
uses them a little differently. Added eslint rule so that we don't
import either of these to app.

It was breaking being able to edit user bio on river, probably among
other things.