Check owner permissions before allowing a user to create a space stream (#3407)
This seems obvious right? If a space exists on the blockchain, but the
stream doesn't yet exist in Casablanca,
we should only let a user with PermissionAddRemoveChannels create it.

Reason being, you don't want random people instantiating your towns, but
looks like we want to be able to designate someone other than yourself
to be the owner when minting a town.

Came accross this running current sdk tests against entitled backend,
was curious why tests were making it so far.