Jt/fix decryption 2 (#3494)
Closes HNT-2521, HNT-2494.

This Pr:
- [x] sends an event correctly to RDE (RiverDecryptionExtension) to
prompt Key Search when timeline first opens in app. Before we were
sending an event, but it was at times not a `message` and additionally
it was missing the `spaceId`.

- [x] fixes fallbackKey retrieval bug that would invariably read oldest
(not newest) key from User stream

After this PR:
- Key seeking starts when channel timeline open (if user does not have
megolm keys on device)
- ToDevice messages continue to work after target recipient refreshes
their deviceKey set (curve25519, fallbackKey).