app: prepopulate channel events (hnt-2417) (#3454)
This is a workaround to pre-populate messages while local persistence
isn't yet implemented (to be used in
https://github.com/HereNotThere/harmony/pull/3452). I guess event when
it gets implemented we would need something similar for the first run.

This is a pretty naive approach, but it's also pretty efficient to keep
it simple. Everything is done in a sequence allowing some pauses to not
cog up network or performance. check constants for thresholds etc.