app: display solo threads (hnt-2573) (#3523)
we shouldn't filter single-user threads, this is a mistake of mine

this could potentially be the issue behind hnt-2573 - we shouldn't close
the issue though

there are clues in the issue that points to this

possible scenario: alice writes a message, replies to her own message
and mentions bob. in this case bob wouldn't see the thread since it was
filtered out.