app: add search to mobile tab bar (hnt-2723 hnt-2734) (#3588)
Prior to this PR message search was intermixed with the results from the
"Jump to..." prompt on the home screen.
There's now a new Search tab, with similar functionality as the "Jump
to" but with broader results (i.e. including messages).
On the home screen we now have "Jump to" focusing on Chanels and Persons
like before, but with a hint to "Search for the term in messages", this
opens up the new search tab.

Since the home screen and search tab share a lot of functionality this
PR also contains a bit of refactor (thus renaming and moving around) in
order to avoid repetition. This could be even tighter but I'm also
afraid of unexpected regressions so I've kept it within sanity for this
pass.


https://github.com/HereNotThere/harmony/assets/557990/f081affa-df7a-43d6-ae52-da92a0081c41