Update the zion client to wait for block confirmation in join channel (#3625)
I think we want this all the time, in my previous PR I added pending
joined and pending invited data structures that we can use to
optimistically update the UI if this ends up taking too long. I also
added an optional parameter to skip the wait if we need that.