Protobufs to support usernames, display names (#3640)
Closes HNT-2771

This sets us up to store encrypted usernames/display names in streams. A
couple of design decisions:

- Since usernames can be unique within a town, it makes sense to use
megolm sessions for encryption. Members will need to decrypt other
member's usernames/display names so it's less onerous to re-use
encryption keys that are scoped for the town.

- With that said, we currently don't create megolm sessions at the town
level, but rather at the channel level. To achieve the above with our
Megolm encryption protocol, we'll need to make a small change to
KeyRequests to request keys upon joining a town scoped to the town, not
any channel therein.

- Since nodes are storing ciphertext for usernames/display names, it's
assumed nodes can't perform uniqueness checks when usernames are edited.
And usernames are needed to be unique per Town. So the client will need
to perform this check when usernames are set/edited during Town join
flow.

- I added an op code in the protocol for default username editing and
placed this message in the UserSettingStream. Additionally, I suffixed
the UserSettingStream snapshot by spaceId. This is more of a redundancy
decision. Users may clear their devices or otherwise lose their megolm
keys. They should be able to reclaim their default usernames and decrypt
them. By encrypting default usernames on write by each spaceId the user
is a member of, if the user loses their keys, they can login and run a
race to retrieve keys from the first member of a space that they are
still a member of lessening the probability that they lose their default
username to the probability that { either the user is no longer a member
of any prior towns that they had when setting their default user name ||
their are no members of those towns online to share keys }.