app: placeholders for town access modal (hnt-2383) (#3705)
far from complete, but works as a first iteration.

main caveats:
- login is handled as a "step" in the modal instead of being fully
refactored into the the "Join Space" button. There's a lot of things
going on in the `LoginComponent` and I think this should be treated
separately
- having 2 routes (one public and one authed) makes the popup disapear
after login in or switching networks. This could be fixed too but
requires moving around auth and routes quite a bit
- some placeholders and all edge cases aren't handled