app: refinements to join flow (#3725)
Hooked up UI to membership data (gating, price). 

@ttillberg refactored the flow a bit: 
- connect wallet from within the join view instead of opening access
modal
- checks membership entitlement _before_ trying to join, and just
removes the join button altogether if user doesn't have the right assets

This largely removes the need for the TownAccess modal altogether - the
only time it should appear now is if you are logged out, then you
connect, then you try to join, then you have to login. We can fix this
in the future but it's ok for now while we're in this intermediate state