app: basic privy embedded wallet flow (#3783)
Converts our login flow to use Privy. App level changes only, lib is not
impacted.

- Sets up the PrivyProvider and uses their PrivyWagmiConnector in place
of WagmiConfig
- App basically has 2 auth states: logged in or logged out. After
logging in with Privy, it should automatically log you in to River,
whether you are a new or returning user. Maybe some kinks to this that
will need to be worked out.
- External wallets no longer apply. I noticed one weird clash with
Privy/Wagmi/Metamask that may only effect local dev so best bet for now
is to DISCONNECT the site from the extension itself.
- converted `useAuth` to a React Context - there's a privy callback for
successful login that should only be called once, and it made sense to
store it there
- Moved a bunch of now unneeded components to `archive` in case we need
to reference them later

Rest of flow remains the same, you still need to fund your new privy
wallet in order to create towns or join them.

Tasks before I merge:
- [x] Log in, create town (local)
- [x] Join town, have conversation (local)
- [x] Log in, create town (app-test-beta)
- [x] Join town, have conversation (app-test-beta)
- [x] Login on mobile (app-test-beta) 
- [x] Add VITE_PRIVY_ID to Render