lib: add setSigner to useWeb3Context (#3804)
Running into some issues with Privy, I think their sdk can take a bit
longer to load the wallet than the Web3Context is initialized, and for
some reason our signer isn't updated properly. Privy has a `ready` prop
which the app is using, but still running into this occasionally 🤷🏻
In the meantime, we can add functionality to set the signer, and then
the app can determine when it's actually ready and update if needed.