Key request improvement HNT-2868 (#3884)
Closes HNT-2868 - This PR changes Key Request protocol to make a plain
text request on the streamId that the solicitor is requesting keys for
in effect demultiplexing N*D requests that were previously being made to
N users and their D devices. Responses still transmit through to device
messages. Support for dm, gdm in a subsequent pr.



todo
- [x] need to test further especially around client restarts

placing in DRAFT until the above is satisfied