app: tweaked privy approach (#3909)
There's a couple things going on with logging in:

1. We're probably running into an issue with wagmi's `useWalletClient` -
basically, the return value is undefined somewhat frequently. It must
have to do with their internal react-query config and `refetchOnFocus`,
as you can open console and suddenly the app appears to log you in
automatically, b/c the wallet client is now populated. I found a couple
issues around this https://github.com/wagmi-dev/wagmi/issues/2784 and
https://github.com/wagmi-dev/references/issues/167#issuecomment-1576756063.
I'm especially inclined to think this is related because at least 1
other person said it only effects them in production build, which is the
primary place we experience it.
So, I moved away from depending on `useWalletClient`, and am polling for
`getWalletClient`, which takes react-query out of the picture and others
mentioned as being more reliable

2. I had a flawed understanding of privy's login.onSucces callback, and
actually the closure there doesn't contain the most updated references.
So post logging in to Privy, you would try to log in to River, but
sometimes no signer was set and you couldn't. I updated so that post
login sets a flag, which then turns on polling for the signer, and then
logs you in.

Going to experiment with this on the render preview