Jt/hnt 3358 stop recursive search for keys (#3917)
Closes HNT-3358

Specifically, this PR eliminates a source of recursion in
RiverDecryptionExtension that is now unnecessary as we are sending
keySolicitations to the target stream that the session was created in,
rather than the member device set. This should alleviate CPU load on the
client and to the node from addEvent calls (will run load tests on this
branch @artem-hnt ).

Next, this PR changes key request to ask for all missing sessionIds
sequentially in one invocation which obviates the need to await a
subsequent call to askStreamForKeys, which now is fully event driven
(net of an initial call when the app loads).