HNT-3361 - Key fulfillment improvement (#4002)
Key fulfillment prior to this PR was not including the sessionIds
actually sent but rather the sessionId requested. The 2 may differ on a
keySolicitation fulfillment depending on the sessions in possession by
the client handling the fulfillment. This changes the session_id field
to be of type `repeated` on Fulfillment message and includes that exact
session_ids so client soliciting keys can re-solicit if they didn't
previously receive the right keys.

Todo:
- [ ] test with 3 clients on sample-app