Update CryptoStore (#4009)
In the spirit of "let's make things easier to read", this PR replaces
the existing calls to IndexedDB from `CryptoStore` with a more modern
ORM, [(Dexie)](http://dexie.org/), which seems very promising.

Two main improvements:
- No more callback-based APIs. Instead we'll now be able to use
async/await everywhere. The code is easier to follow.
- No more need for a separate database implementation for testing, as we
can rely on `fake-indexeddb`.

We also get versioning for free, which will close
[HNT-3342](https://linear.app/hnt-labs/issue/HNT-3342/implement-idx-versioning).
This PR intentionally tries to preserve all logic exactly as-is. 1 step
at a time.

### Note
I found an issue (known?) in the conversation tests
(`aliceAndFriends**.test.ts`) — turns out that they are **not** actually
end-to-end, but instead depend on all clients reading from the same
database in local storage.
 No `toDeviceMessages` are parsed during the conversation tests.
This needs attention, especially as we're getting ready to phase out the
RiverDecryptionExtension. [Added a
ticket.](https://linear.app/hnt-labs/issue/HNT-3569/make-sure-that-the-conversation-tests-in-the-sdk-send-olm-messages)