set deploy timeouts to 5 minutes instead of 30 (#3999)
when we deploy a new container, aws code-deploy brings it up as "green"
group in our "blue-green" setup. our alb sends a request to its /info
endpoint as a healthcheck. if the request is successful, green gets
promoted to blue. the issue is that our /info healthcheck doesn't run
any meaningful checks. so it's actually very easy for us to deploy a bad
container that will pass the "healthcheck", replace "blue", and even
sometimes acquire a uuid db-lock which will not only take-over http
traffic from blue, but also crash-exit it, making it even harder to
execute a seamless rollback (since blue not only needs to get
re-promoted, but actually rebooted). all before "green" crash-exits
itself.

so by the time our bad "green" is promoted, "blue" is down, and
eventually "green" is also down. so we're down. "luckily" aws
code-deploy detects this, and rolls back the deployment. so "blue" is
back up, "green" is terminated, and everything is well. unfortunately,
aws code-deploy does not stop there. it will keep trying to make the
deployment work. so the whole cycle repeats again.

a few additional issues here: aws-code deploy will not register these
cycles of failures as a failed deployment until the 30 minute timeout.
so we won't get a failure notification until after 30 minutes have
passed (the default timeout). and since the river node keeps going up
and down in quick succession, our uptime monitors don't always detect it
as they are set to ping the river node every 30 minutes. this is due to
how expensive datadog uptime monitors get when you run them from
multiple geographical regions (we run them from 10 different locations).
as an additional measure, I created a separate uptime monitor that runs
1/min from a single geographical region, which should improve our uptime
monitors and alerts.

back to code-deploy and how to improve it: the lowest hanging fruit here
is introducing a timeout to our deploys. unless it's successful in 5
minutes, aws-code deploy should just stop. that's what this PR does.
additionally, i think code-deploy should only attempt to deploy once. if
the first attempt fails, we should just call it and fail the deployment.
i'm looking into configuring code-deploy for this.

but the thing that trumps everything is proper health checks. river-node
should not signal that it's ready to serve public http traffic, before
it knows that it's actually ready. that will have the greatest impact on
uptime during cd. it will have a significant impact on how we
continuously deploy to production. until then, we will keep going down
in the likely event of deploying a bad container.