Kerem/hnt 3538 every pr should create its own custom dns record (#4024)
This PR is our introduction to transient environments. It lays the
groundwork, and introduces our first transient environment element: The
render preview DNS records. When a PR is created, a dedicated Cloudflare
DNS record, referencing the pr # is created. Upon closing the PR (by
merging or by closing), the environment is deprovisioned. Upcoming work
in the transient environments projects will follow this framework.