Only getStream for room device list if creating a new session (#4063)
Closes HNT-2612 - Only call getStream on the critical path of
sendMessage if creating a new session as the side effect of creating a
new session is sharing it with the device list of the current members
list of the stream. This PR also removes some unused code from
`megolm.ts`. Tested on the sample-app locally with 2 clients. A follow
on PR will store megolm sessions in IDB, such that sessions are
preferred to be loaded from local storage. This will further reduce new
session creation and therefore the frequency with which a client calls
getStream for the entire member list.