New wallet linking address (#4086)
When we update contracts we should run start local dev locally (or just
`Start Blockchains & Deploy Smart Contracts`) and check in the json
files

We do this because we include the json files in the build (yarn), but
afik we need to run yarn to build the contracts. Something has to give,
this is the comprimise we came up with.