disabling env resets on weekends (#4048)
We received "@ channel" notifications over the weekend as the cron job
kept resetting the environment. This PR disables env resets over the
weekend. Unintended benefit: we get to let the environment run a bit
longer over the weekend, until the Monday morning reset, so we get a
chance to observe behavior during longer runs.