updating the rpc proxy version for aggressive logging and better debugs (#4111)
I just updated the proxy worker version. This one has a lot more console
logs. I also pulled the provider settings to a more visible spot so it's
easy for everyone to change them at will.

If you want to read about the exact changes in this patch, here's the
changelog:
https://github.com/whats-good/nexus/blob/main/packages/nexus/CHANGELOG.md

You can take a look at this more detailed documentation page if you want
to learn about the inner workings of the proxy cloudflare worker:
https://nexus.whatsgood.dog/recipes/cloudflare-worker