river/web3: single ISpaceDapp compatible with v3 + v4 (#4100)
ISpaceDapp should be the final duplicated file between v3 and v4, and
now it's updated to handle the different versions correctly.

Everything is still set to v3, so there should be no breaking changes
anywhere as far as types or functionality.

There is now a Versions, defaultVersion, and TypeDefaultVersion that
should be determining the correct types and/or logic across all the
shimmed space dapp logic. I'm hoping that the end goal here _should_ be
just to modify defaultVersion to 'v4', and then have minimal updates in
lib/app/sdk. I think this can set us up well for future versioning as
well