HNT-3661 keyInitialization fixes (#4131)
Closes HNT-3661

Several issues were breaking keyInitialization:

1. We were calling `import_session` to create an Inbound group session
from a session sent during key initialization, however the session key
was not in export format and so this was failing to actually write the
session to IDB and re-attempt decrypting the event with
OLM.BAD_SESSION_KEY exception.
2. We were sending keys to ourself which was causing toDevice exceptions
for every keyInitialization. This was possibly a red herring for the
first issue, but it caused consistent noise in the console and so I
stopped keyInitializing to our user's devices (for the time being).
3. We were retrieving the deviceKey from in-memory storage when sending
toDevice messages that was inconsistent with the olm encrypted envelope.
This was causing decryption errors on receipt. We need to clean up all
these in-memory storage data structures, which will be the subject of
another PR.

Note there are 2 todos inline that i will address with a level head
tomorrow. I don't think they block this PR. I've tested this on 3
clients each with it's own browser context testing keyInitialization and
keySolicitation on the sample-app. Both work with this PR.