river/web3: add viem compatible PioneerNFT (#4130)
didn't realize this one was in there too.

When testing, I discovered `const mintReward = await
this.pioneerNFTShim.read.getMintReward()` in
`PioneerNft.getContractState()` throws an error - guess the abi or
typings are outdated? It's reproducible on
https://sample-app-b0wc.onrender.com/alpha-access and I assume it's been
that way for a while, I guess no one uses this anymore. Do we still need
this page? I can remove and remove the method as well.