app: prevent update twice + more verbose logging (hnt-3333) (#4160)
I managed to reproduce the issue where the update prompt appears twice
in a row. It can happen if the the reload happens while a new update is
installing. I guess this may happen if you open the app very shortly,
just enough time to trigger installation without activating it
explicitly (activation not being possible because the app is open in
another tab - or differently on mobile where the app isn't actually
"closed")

To mitigate this, and for the purpose of logging, this PR will always
show the current version number for 5 seconds. Followed by reload prompt
if necessary. This way, in the case of twos consecutive updates, the
second will have the time to install over the preceding one before the
prompt appears.

In case this experiment fails, it will nevertheless be easier to figure
if an update actually happened as we will always be able to verify the
previous version hash before hitting `reload`

On the side:

- I've removed the startup timeout in favour of a shorter interval. On
browser the check is always done instantly on startup and on mobile the
component is still mounted when opening the app.
- more verbose logging. 
Another issue we have noticed is the that a new version won't update if
the app is opened in several other tabs (notably if it got into a reload
loop). In that case hitting reload will keep trying to install the same
version over and over. This will be more noticeable now by showing the
version hash. I've also added logging for all occurrences of
`location.reload`. More over logging is more verbose.