KeyInit to user's own devices (#4167)
With this PR, we initialize keys to user's own devices. It turns out the
error I was seeing yesterday was a red herring due to inconsistent
toDevice session management which fixed in
https://github.com/HereNotThere/harmony/pull/4151.

I also reduced max lead wait time to ask for keys (per event basis) to 2
seconds, which should make decryption of the timeline initially complete
in roughly 1/2 the time.

There is a race between keyInitialization and keySolicitation as a
result of lowering the lead time to ask for keys. With IDB transactions,
I wonder if we can detect if a lock is held to write to an IDB table,
which would occur during key import for the initialized key that is sent
with a new session. @erikolsson thoughts ?

Tested locally between 2 sessions 1 on brave, 1 on chrome for the same
User. Created a town sent a message. It decrypted immediately.
Responded. Worked. Then performed a negative test without this commit
and replicated the problem decrypting between 2 devices for the same
user chatting in a channel.