user feedback -> linear (#4173)
This PR uses the existing gateway endpoint for `user-feedback`, and
pushes error reports directly to linear. this is primarily done to avoid
missing user feedback due to ad blockers. it also optimistically
includes the datadog session url. if it wasn't adblocked, the url will
show the user session that led to the feedback. otherwise it will just
display an empty datadog dashboard. Here's a screenshot of an
auto-captured user feedback:


![image](https://github.com/HereNotThere/harmony/assets/5973785/8b6f67d2-00e4-42c3-b572-5e1348dfb343)