use unique database for each client during tests (#4146)
A lot of learnings splitting up the IDB — not only did it clearly
surface the need for transactions, but also how some of the tests needed
to be changed. Stacking #4147 on top of this one. I would like to show
which tests are breaking, and exactly which steps we're taking to fix
them as opposed to committing one big PR.