app: config - remove define process.env (#4185)
Even if the previous version with `NODE_ENV: stringify(mode)` seemed to
be correct( it was passing "production" as NODE_ENV to the `console.log`
in the SW) it didn't compile workbox in production mode. Removing the
`define` all together seems to do the trick.

In any case the hack seems to be obsolete, all tests pass and the app
builds and runs without it.