Fix out of order threads (#4271)
This was my oversite, for a long time.

When threaded messages get loaded from scrollback, the encrypted event
will get properly prepended to the timeline, but the threadId is
encrypted. When we decrypt, we just “update” the event and replace it in
the timeline, which works for the main channel, but the thread bit was
new and we were just appending them always. Because events can decrypt
in any order… we have to sort the threads when we add to them. Sorry
this took so long to track down.

fixes HNT-3798