Remove the unused key request data structures (#4289)
We want to move to device messages to their own stream and properly
snapshot and expire them.

This removes the “ToDeviceOp” that makes the payload a little harder to
split by device id, and removes the “request” kind in the encrypted
payload since both were unused

I had to re-configure the toDeviceMessage.test.ts tests. They were using
key request payloads to test decryption, and they started failing when I
switched to key responses because they riverDecryptionExtensions kicked
in and started decrypting them, and we can’t decrypt a toDeviceMessage
twice. I added a new event that i can listen to in the test.