keySolicitation race fixes on client startup (#4225)
We currently start the RDE on client startup and then initializeUser
which initializes as part of the routine all joined streams prepending
events prior to the snapshot, and appending events after the latest
snapshot. KeySolicitation events and Fulfillment events are added in
order, which means there's a race that ensues to process KeySol events
prior to Fulfillments. This PR fixes that race.