use-timeline-store optimization (#4339)
This wasn’t on my list to do, but this has bugged me for a long time and
I’m a bit delerious right now and I squinted at the code and it made
sense to change it.

Before, if we had a list of events, we would call setState on the
timelineStore for every single event. Now we call setState once and do
all the operations for all of events inside the setstate callback.

zustand was probably doing something to coalesce them, but why make it
work when we can just make a single update.