cleanup: remove code for non-prekey messages (hnt-3096) (#4327)
This PR removes support for non-prekey messages. We don't support
sending them, so we shouldn't support decrypting them. Also cleans up
some hard-to-follow Matrix code. If we need support for non-prekey
messages in the future, we can write something much nicer and easy to
read. This should close
[HNT-3096](https://linear.app/hnt-labs/issue/HNT-3096/remove-dead-code-paths)
since we now have minimal amounts of unused code remaining from the
Matrix libs.

Also changes the API a bit, now that we have the `OlmMessage` proto
generated class, let's use that instead of `IEventOlmDecryptionResult`