app: update privyReady check, move reloadPrompt to Main.tsx, remove lazy DebugBar (#4334)
1. Removed my bugged `showWelcomeLayout` value that flip flopped. Erik
addressed this in https://github.com/HereNotThere/harmony/pull/4180. I
don't think the check is necessary anymore b/c other updates to sign in
flow were made so it's ok not to check the status before
ZionContextProvider is instantiated. I added a simpler check in
AllRoutes
2. Removed lazy DebugBar import. After removing the check, i noticed
that multiple clients were back. And found that the same thing appears
in alpha.towns.com. I have the feeling that we've always had multiple
clients on startup. I don't understand why, but it seems like the lazy
import of DebugBar is causing this. Removing the lazy import, I get only
a single zion client. I'm not sure how Erik's fix played into getting a
single client, but I guess it was some race condition.
3. I moved the ReloadPrompt to Main.tsx as a follow up to
https://github.com/HereNotThere/harmony/pull/4248. Multiple clients
indicates that multiple reload prompt/service worker registrations might
have been going on, and if we've had multiple clients since alpha maybe
that explains some of the buggy behavior. We could keep this in App.tsx
but it has no deps on anything so moving to Main seems fine.