Fix edge case in confirmed message updates (#4348)
I added message confirmation updates yesterday, but it triggered an edge
case that I wanted to go ahead and fix this morning.

We have two layers of things going on, there is a single timeline in the
sdk of all events, and there is a “combined” timeline in the react sdk
of all the updates/edits/deletes/etc.

When we edit a message, we just append a new event to the timeline on
the node and in the sdk. The react sdk (lib) recognizes that the second
event is an edit, and replaces the first event in it’s data store.

This gets really complicated with encryption, where decrypted messages
replace their previously encrypted events, but also do edit/redact logic
based on the decrypted properties.

My previous change added another update to the system, which could cause
a previosly edited message to get swapped out by the confirmation of the
original event.

I fixed this to add a special update for confirmations that only
modifies the confirmedEventNum and confirmedInBlockNum properties. A
sniper’s approach if you will.

Updated tests and spot checked test output.