creating conditional logic for render previews (#4390)
This PR creates a new mode for transient environments: `lite`. A lite
transient env will only provision the Cloudflare dns records so the
render apps continue functioning with privy.

Will sunset the label `provision transient environment` in exchange of
`provision full transient environment`. If this label is not set, the
lite env will be created by default, which will be just the 2 dns
records for the web app and the sample app. if the `provision full
transient environment` label is selected, then we'll create the rest of
the stack.

I had to introduce an additional component to make this work. Basically
there's a step in our web app builds, which sets the csb url of the
render app based on the PR number. We now also need to get the pr
labels, so render needs a github token to read the pr label. I've
created a token and added it to render build env vars. Since this env
vars is not prefixed with `VITE_`, it won't make it to the static build,
which is good. It will only exist during build time within render. We
require this step, because we want `transient lite` envs to connect to
the original river node

Context:

<img width="1365" alt="image"
src="https://github.com/HereNotThere/harmony/assets/5973785/1066509b-8f9e-43a0-a4fe-f96d687aafb0">