Enable the bobFlushes presync test (#4416)
In the massive pr I fixed a bug where we were just looking for the
minipool events in the sync response, and ignoring events that had been
committed to blocks. If there was enough time to make a block, the test
would fail. The new code looks like

https://github.com/HereNotThere/harmony/pull/4296

```
const spaceResponse = await bob.getStream({ streamId: spacedStreamId })
    const envelopes = [
        ...unpackStreamResponse(spaceResponse).miniblocks.flatMap((x) => x.events),
        ...unpackEnvelopes(spaceResponse.stream!.events),
    ]

```

and this test seems to work every time now.